Slot Sounds, Logins, Metamasks, OH MY!
Implemented Sounds for Slots spinning, debating on whether to add a jackpot sound effect or not, i will still be adding a sound effect for each game, but those will be in separate commits/merges.

I also cheated by making a fake animation by having the symbols iterate through the array of emojis randomly, creating the illusion of a spinning wheel.

Also implemented (per Andrews request) a login screen with instructions on how to connect ones wallet to the website.

I have also fixed the issue with the blackjack dealer card, by updating the css and the renderHand function